+++ Kernel/Modules/AgentTicketProcess.pm
@@ -509,7 +509,8 @@ sub _RenderAjax {
     DYNAMICFIELD:
     for my $DynamicField ( sort keys %DynamicFieldValues ) {
         next DYNAMICFIELD if !$DynamicField;
-        next DYNAMICFIELD if !$DynamicFieldValues{$DynamicField};
+        next DYNAMICFIELD if !defined $DynamicFieldValues{$DynamicField};
+        next DYNAMICFIELD if !length $DynamicFieldValues{$DynamicField};
 
         $DynamicFieldCheckParam{ 'DynamicField_' . $DynamicField } = $DynamicFieldValues{$DynamicField};
     }
@@ -969,7 +970,7 @@ sub _GetParam {
             # If we had neighter submitted nor ticket param get the ActivityDialog's default Value
             # next out if it was successful
             $Value = $ActivityDialog->{Fields}{$CurrentField}{DefaultValue};
-            if ($Value) {
+            if ( defined $Value ) {
                 $GetParam{$CurrentField} = $Value;
                 next CURRENTFIELD;
             }
@@ -977,7 +978,7 @@ sub _GetParam {
             # If we had no submitted, ticket or ActivityDialog default value
             # use the DynamicField's default value and next out
             $Value = $DynamicFieldConfig->{Config}{DefaultValue};
-            if ($Value) {
+            if ( defined $Value ) {
                 $GetParam{$CurrentField} = $Value;
                 next CURRENTFIELD;
             }
@@ -4162,7 +4163,7 @@ sub _StoreActivityDialog {
                 # if we have an invisible field, use config's default value
                 if ( $ActivityDialog->{Fields}{$CurrentField}{Display} == 0 ) {
                     $TicketParam{$CurrentField} = $ActivityDialog->{Fields}{$CurrentField}{DefaultValue}
-                        || '';
+                        // '';
                 }
 
                 # only validate visible fields